jupiter_juno_robot/requirements.txt
# Core ROS 2 dependencies (should be installed with ROS 2)
# rclpy
# cv_bridge

# Computer Vision and Face Detection
opencv-python>=4.5.0
dlib>=19.22.0
scipy>=1.7.0
numpy>=1.21.0

# Audio and TTS
pyttsx3>=2.90
gtts>=2.2.0
pygame>=2.0.0
simpleaudio>=1.0.4
pyaudio>=0.2.11

# Speech Recognition
SpeechRecognition>=3.8.1

# AI/ML APIs
google-generativeai>=0.3.0
openai>=0.27.0

# Utilities
PyYAML>=6.0
pillow>=9.0.0

# Optional dependencies for better performance
# pocketsphinx>=0.1.15  # For offline speech recognition 